+++ app/app.go
@@ -1,4 +1,4 @@
-package cli
+package app
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -58,7 +58,7 @@ import (
 // Basic utility info
 const (
 	APP  = "Atlassian Cloud Backuper"
-	VER  = "0.0.2"
+	VER  = "0.1.0"
 	DESC = "Tool for backuping Atlassian cloud services (Jira and Confluence)"
 )
 
@@ -221,31 +221,17 @@ func addExtraOptions(m options.Map) {
 	}
 
 	knfu.AddOptions(m,
-		ACCESS_ACCOUNT,
-		ACCESS_EMAIL,
-		ACCESS_API_KEY,
+		ACCESS_ACCOUNT, ACCESS_EMAIL, ACCESS_API_KEY,
 		STORAGE_TYPE,
-		STORAGE_FS_PATH,
-		STORAGE_FS_MODE,
-		STORAGE_SFTP_HOST,
-		STORAGE_SFTP_USER,
-		STORAGE_SFTP_KEY,
-		STORAGE_SFTP_PATH,
-		STORAGE_SFTP_MODE,
-		STORAGE_S3_HOST,
-		STORAGE_S3_ACCESS_KEY,
-		STORAGE_S3_SECRET_KEY,
-		STORAGE_S3_BUCKET,
-		STORAGE_S3_PATH,
-		JIRA_OUTPUT_FILE,
-		JIRA_INCLUDE_ATTACHMENTS,
-		JIRA_CLOUD_FORMAT,
-		CONFLUENCE_OUTPUT_FILE,
-		CONFLUENCE_INCLUDE_ATTACHMENTS,
-		CONFLUENCE_CLOUD_FORMAT,
+		STORAGE_FS_PATH, STORAGE_FS_MODE,
+		STORAGE_SFTP_HOST, STORAGE_SFTP_USER, STORAGE_SFTP_KEY,
+		STORAGE_SFTP_PATH, STORAGE_SFTP_MODE,
+		STORAGE_S3_HOST, STORAGE_S3_ACCESS_KEY, STORAGE_S3_SECRET_KEY,
+		STORAGE_S3_BUCKET, STORAGE_S3_PATH,
+		JIRA_OUTPUT_FILE, JIRA_INCLUDE_ATTACHMENTS, JIRA_CLOUD_FORMAT,
+		CONFLUENCE_OUTPUT_FILE, CONFLUENCE_INCLUDE_ATTACHMENTS, CONFLUENCE_CLOUD_FORMAT,
 		TEMP_DIR,
-		LOG_FORMAT,
-		LOG_LEVEL,
+		LOG_FORMAT, LOG_LEVEL,
 	)
 }
